[refactor] Extract expr code generation to separate _expr module. [python/assassyn/codegen/verilog/_expr/__init__.py] Create new module with main codegen_expr dispatcher function to handle all expression code generation. [python/assassyn/codegen/verilog/_expr/arith.py] Extract arithmetic and logical operations (BinaryOp, UnaryOp, Slice, Concat, Cast, Select, Select1Hot) to dedicated module. [python/assassyn/codegen/verilog/_expr/array.py] Extract array and FIFO operations (ArrayRead, ArrayWrite, FIFOPush, FIFOPop) to dedicated module. [python/assassyn/codegen/verilog/_expr/call.py] Extract call operations (AsyncCall, Bind, WireAssign, WireRead) to dedicated module. [python/assassyn/codegen/verilog/_expr/intrinsics.py] Extract intrinsic operations (PureIntrinsic, Intrinsic, Log) to dedicated module. [python/assassyn/codegen/verilog/design.py] Update import to use _expr module instead of expr module. [python/assassyn/codegen/verilog/expr.py] Remove old monolithic expr.py file. This refactoring improves code organization by breaking down the large expr.py file (528 lines) into smaller, focused modules similar to the simulator _expr structure, making the codebase more maintainable and easier to navigate.
